Relocate get-run-id utility func to cstar.orch.utils
