+++ cstar/orchestration/orchestration.py
@@ -1,7 +1,6 @@
 import asyncio
 import os
 import typing as t
-from datetime import datetime, timezone
 from enum import IntEnum, StrEnum, auto
 from pathlib import Path
 
@@ -672,24 +671,3 @@ def configure_environment(
 
     if run_id:
         os.environ[ENV_CSTAR_ORCH_RUNID] = slugify(run_id)
-
-
-def get_run_id(run_id: str = "") -> str:
-    """Retrieve the current run-id.
-
-    Parameters
-    ----------
-    run_id : str | None
-        If non-null, used to override a pre-configured run id.
-
-    Generate a new run-id if not found in the environment.
-
-    Returns
-    -------
-    str
-    """
-    if run_id:
-        return run_id
-
-    new_run_id = datetime.now(timezone.utc).strftime("%Y%m%d_%H%M%S")
-    return os.getenv(ENV_CSTAR_ORCH_RUNID, new_run_id)
